requirements.txt
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
et-xmlfile==1.1.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.2
openpyxl==3.0.10
pandas==1.4.3
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
urllib3==1.26.11
Werkzeug==2.2.2
xlrd==2.0.1
zipp==3.8.1
